+++ src/colors/syntax/html.colors.ts
@@ -3,14 +3,19 @@ import { SyntaxColors } from '../../types/colors-types';
 const htmlColors: SyntaxColors = (tokens) => {
   return [
     {
-      name: 'Attribute',
-      scope: ['meta.tag', 'meta.tag.inline.any'],
-      settings: tokens.tag.attribute
+      name: 'HTML: Tag',
+      scope: 'entity.name.tag.html',
+      settings: tokens.html.tag.name
     },
     {
-      name: 'Value',
+      name: 'HTML: Attribute',
+      scope: ['meta.tag', 'meta.tag.inline.any'],
+      settings: tokens.html.tag.attribute
+    },
+    {
+      name: 'HTML: Value',
       scope: 'string.quoted.double.html',
-      settings: tokens.tag.value
+      settings: tokens.html.tag.value
     },
     {
       name: 'Text',